requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
pymysql
flask_cors

